requirements.txt
google-generativeai>=0.3.0
typing-extensions>=4.5.0
pytest>=7.4.0
pytest-asyncio>=0.21.1
aiohttp>=3.9.0
flask>=3.0.0
requests==2.31.0
beautifulsoup4
html2text
shortuuid
anthropic
SQLAlchemy
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
PyJWT>=2.8.0
email-validator>=2.1.0.post1
